Remove redundant same-type casts crate-wide
Applied `cargo clippy --fix -- -A clippy::all -W clippy::unnecessary_cast`:
removes casts where the pre-cast expression's type, per rustc's own
inference, is already identical to the cast target (e.g.
`size_of::<T>() as usize as isize` -> `size_of::<T>() as isize`, since
size_of() already returns usize). This is a machine-verified proof of
no-op-ness, not a heuristic — clippy only flags same-type casts, so it
cannot touch any cast that narrows, widens, or changes signedness
(confirmed: the 17 deliberate `as int16_t as uint16_t` float-narrowing-fix
casts from fix-float-narrowing.py, the char-narrowing casts, and the
signedness-changing `as usize as isize` occurrences are all untouched,
since none of those are same-type casts).

743 occurrences across 42 files.

+++ rust/src/dep/extern/json.rs
@@ -236,7 +236,7 @@ unsafe extern "C" fn new_value(
             }
             1 => {
                 if !((*value).u.object.length == 0 as ::core::ffi::c_uint) {
-                    values_size = (::core::mem::size_of::<json_object_entry>() as usize)
+                    values_size = ::core::mem::size_of::<json_object_entry>()
                         .wrapping_mul((*value).u.object.length as usize)
                         as ::core::ffi::c_int;
                     (*value).u.object.values = json_alloc(
@@ -629,8 +629,8 @@ pub unsafe extern "C" fn json_parse_ex(
                                                 | uchar2 & 0x3ff as json_uchar;
                                         }
                                     }
-                                    if ::core::mem::size_of::<::core::ffi::c_char>() as usize
-                                        >= ::core::mem::size_of::<json_uchar>() as usize
+                                    if ::core::mem::size_of::<::core::ffi::c_char>()
+                                        >= ::core::mem::size_of::<json_uchar>()
                                         || uchar <= 0x7f as json_uchar
                                     {
                                         if state.first_pass == 0 {
