+++ rust/src/lib/consolidate/otl/gsub_single.rs
@@ -1748,7 +1748,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                     {
                         s = ((*(*(*h).hh.tbl).buckets.offset(_hf_bkt as isize)).hh_head
                             as *mut ::core::ffi::c_char)
-                            .offset(-((*(*h).hh.tbl).hho as isize))
+                            .offset(-(*(*h).hh.tbl).hho)
                             as *mut ::core::ffi::c_void
                             as *mut gsub_single_map_hash
                             as *mut gsub_single_map_hash;
@@ -1758,7 +1758,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                     while !s.is_null() {
                         if (*s).hh.hashv == _hf_hashv
                             && (*s).hh.keylen as usize
-                                == ::core::mem::size_of::<::core::ffi::c_int>() as usize
+                                == ::core::mem::size_of::<::core::ffi::c_int>()
                         {
                             if memcmp(
                                 (*s).hh.key,
@@ -1771,7 +1771,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                         }
                         if !(*s).hh.hh_next.is_null() {
                             s = ((*s).hh.hh_next as *mut ::core::ffi::c_char)
-                                .offset(-((*(*h).hh.tbl).hho as isize))
+                                .offset(-(*(*h).hh.tbl).hho)
                                 as *mut ::core::ffi::c_void
                                 as *mut gsub_single_map_hash
                                 as *mut gsub_single_map_hash;
@@ -2130,7 +2130,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                     (*s).hh.tbl = (*h).hh.tbl;
                     (*s).hh.next = NULL;
                     (*s).hh.prev = ((*(*h).hh.tbl).tail as *mut ::core::ffi::c_char)
-                        .offset(-((*(*h).hh.tbl).hho as isize))
+                        .offset(-(*(*h).hh.tbl).hho)
                         as *mut ::core::ffi::c_void;
                     (*(*(*h).hh.tbl).tail).next = s as *mut ::core::ffi::c_void;
                     (*(*h).hh.tbl).tail = &raw mut (*s).hh as *mut UT_hash_handle;
@@ -2285,7 +2285,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                     _hs_psize = _hs_psize.wrapping_add(1);
                     _hs_q = (if !(*_hs_q).next.is_null() {
                         ((*_hs_q).next as *mut ::core::ffi::c_char)
-                            .offset((*(*h).hh.tbl).hho as isize)
+                            .offset((*(*h).hh.tbl).hho)
                             as *mut UT_hash_handle
                     } else {
                         ::core::ptr::null_mut::<UT_hash_handle>()
@@ -2303,7 +2303,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                         _hs_e = _hs_q;
                         _hs_q = (if !(*_hs_q).next.is_null() {
                             ((*_hs_q).next as *mut ::core::ffi::c_char)
-                                .offset((*(*h).hh.tbl).hho as isize)
+                                .offset((*(*h).hh.tbl).hho)
                                 as *mut UT_hash_handle
                         } else {
                             ::core::ptr::null_mut::<UT_hash_handle>()
@@ -2314,7 +2314,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                         if !_hs_p.is_null() {
                             _hs_p = (if !(*_hs_p).next.is_null() {
                                 ((*_hs_p).next as *mut ::core::ffi::c_char)
-                                    .offset((*(*h).hh.tbl).hho as isize)
+                                    .offset((*(*h).hh.tbl).hho)
                                     as *mut UT_hash_handle
                             } else {
                                 ::core::ptr::null_mut::<UT_hash_handle>()
@@ -2322,10 +2322,10 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                         }
                         _hs_psize = _hs_psize.wrapping_sub(1);
                     } else if by_from_id(
-                        (_hs_p as *mut ::core::ffi::c_char).offset(-((*(*h).hh.tbl).hho as isize))
+                        (_hs_p as *mut ::core::ffi::c_char).offset(-(*(*h).hh.tbl).hho)
                             as *mut ::core::ffi::c_void
                             as *mut gsub_single_map_hash,
-                        (_hs_q as *mut ::core::ffi::c_char).offset(-((*(*h).hh.tbl).hho as isize))
+                        (_hs_q as *mut ::core::ffi::c_char).offset(-(*(*h).hh.tbl).hho)
                             as *mut ::core::ffi::c_void
                             as *mut gsub_single_map_hash,
                     ) <= 0 as ::core::ffi::c_int
@@ -2334,7 +2334,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                         if !_hs_p.is_null() {
                             _hs_p = (if !(*_hs_p).next.is_null() {
                                 ((*_hs_p).next as *mut ::core::ffi::c_char)
-                                    .offset((*(*h).hh.tbl).hho as isize)
+                                    .offset((*(*h).hh.tbl).hho)
                                     as *mut UT_hash_handle
                             } else {
                                 ::core::ptr::null_mut::<UT_hash_handle>()
@@ -2345,7 +2345,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                         _hs_e = _hs_q;
                         _hs_q = (if !(*_hs_q).next.is_null() {
                             ((*_hs_q).next as *mut ::core::ffi::c_char)
-                                .offset((*(*h).hh.tbl).hho as isize)
+                                .offset((*(*h).hh.tbl).hho)
                                 as *mut UT_hash_handle
                         } else {
                             ::core::ptr::null_mut::<UT_hash_handle>()
@@ -2355,7 +2355,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                     if !_hs_tail.is_null() {
                         (*_hs_tail).next = if !_hs_e.is_null() {
                             (_hs_e as *mut ::core::ffi::c_char)
-                                .offset(-((*(*h).hh.tbl).hho as isize))
+                                .offset(-(*(*h).hh.tbl).hho)
                                 as *mut ::core::ffi::c_void
                         } else {
                             NULL
@@ -2366,7 +2366,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
                     if !_hs_e.is_null() {
                         (*_hs_e).prev = if !_hs_tail.is_null() {
                             (_hs_tail as *mut ::core::ffi::c_char)
-                                .offset(-((*(*h).hh.tbl).hho as isize))
+                                .offset(-(*(*h).hh.tbl).hho)
                                 as *mut ::core::ffi::c_void
                         } else {
                             NULL
@@ -2382,7 +2382,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
             if _hs_nmerges <= 1 as ::core::ffi::c_uint {
                 _hs_looping = 0 as ::core::ffi::c_uint;
                 (*(*h).hh.tbl).tail = _hs_tail;
-                h = (_hs_list as *mut ::core::ffi::c_char).offset(-((*(*h).hh.tbl).hho as isize))
+                h = (_hs_list as *mut ::core::ffi::c_char).offset(-(*(*h).hh.tbl).hho)
                     as *mut ::core::ffi::c_void as *mut gsub_single_map_hash
                     as *mut gsub_single_map_hash;
             }
@@ -2443,13 +2443,13 @@ pub unsafe extern "C" fn consolidate_gsub_single(
             let mut _hd_bkt: ::core::ffi::c_uint = 0;
             if _hd_hh_del == (*(*h).hh.tbl).tail {
                 (*(*h).hh.tbl).tail = ((*_hd_hh_del).prev as *mut ::core::ffi::c_char)
-                    .offset((*(*h).hh.tbl).hho as isize)
+                    .offset((*(*h).hh.tbl).hho)
                     as *mut UT_hash_handle
                     as *mut UT_hash_handle;
             }
             if !(*_hd_hh_del).prev.is_null() {
                 let ref mut fresh0 = (*(((*_hd_hh_del).prev as *mut ::core::ffi::c_char)
-                    .offset((*(*h).hh.tbl).hho as isize)
+                    .offset((*(*h).hh.tbl).hho)
                     as *mut UT_hash_handle))
                     .next;
                 *fresh0 = (*_hd_hh_del).next;
@@ -2458,7 +2458,7 @@ pub unsafe extern "C" fn consolidate_gsub_single(
             }
             if !(*_hd_hh_del).next.is_null() {
                 let ref mut fresh1 = (*(((*_hd_hh_del).next as *mut ::core::ffi::c_char)
-                    .offset((*(*h).hh.tbl).hho as isize)
+                    .offset((*(*h).hh.tbl).hho)
                     as *mut UT_hash_handle))
                     .prev;
                 *fresh1 = (*_hd_hh_del).prev;
